% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.osrm.school.drive-package.R
\docType{package}
\name{dc.osrm.school.drive-package}
\alias{dc.osrm.school.drive}
\alias{dc.osrm.school.drive-package}
\title{dc.osrm.school.drive: Social Data Commons: Education & Training: School Drive Times}
\description{
Allows user to easily get drive time to Virginia post-HS education data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
